+++ nym-vpn-core/crates/nym-ipc/src/server.rs
@@ -7,31 +7,31 @@ use tokio::io::{AsyncRead, AsyncWrite};
 use tokio_stream::Stream;
 use tonic::transport::server::Connected;
 
+use crate::authentication::AuthenticationMaterial;
+
 #[allow(unused)]
 pub async fn create_incoming(
     socket_path: std::path::PathBuf,
-    #[cfg(target_os = "windows")] nym_certificate_serial_number: String,
-    #[cfg(target_os = "macos")] signing_requirement: String,
-    #[cfg(unix)] shutdown_token: tokio_util::sync::CancellationToken,
+    auth_material: AuthenticationMaterial,
 ) -> Result<impl Stream<Item = Result<impl AsyncRead + AsyncWrite + Connected + 'static>>> {
     #[cfg(target_os = "macos")]
     {
         #[cfg(all(debug_assertions, not(feature = "xpc")))]
         {
-            crate::uds::incoming(socket_path, shutdown_token).await
+            crate::uds::incoming(socket_path, auth_material).await
         }
         #[cfg(any(not(debug_assertions), feature = "xpc"))]
         {
-            crate::xpc::incoming(signing_requirement, shutdown_token)
+            crate::xpc::incoming(auth_material)
         }
     }
     #[cfg(target_os = "linux")]
     {
-        crate::uds::incoming(socket_path, shutdown_token).await
+        crate::uds::incoming(socket_path, auth_material).await
     }
 
     #[cfg(target_os = "windows")]
     {
-        crate::named_pipe::incoming(socket_path.into_os_string(), nym_certificate_serial_number)
+        crate::named_pipe::incoming(socket_path.into_os_string(), auth_material)
     }
 }